keep splitting things
