+++ src/utils/rgb1to256.ts
@@ -1,0 +1,3 @@
+export function rgb1to256(value) {
+  return Math.round(parseFloat(value.toFixed(4)) * 255);
+}
